+++ src/AppserverIo/Appserver/Core/Api/Node/AnalyticNode.php
@@ -20,6 +20,8 @@
 
 namespace AppserverIo\Appserver\Core\Api\Node;
 
+use AppserverIo\Description\Api\Node\AbstractNode;
+
 /**
  * Node which represents a collection of analytic steps run on a certain URI.
  *