+++ src/AppserverIo/Appserver/Core/Api/Node/FormErrorPageNode.php
@@ -20,6 +20,7 @@
 
 namespace AppserverIo\Appserver\Core\Api\Node;
 
+use AppserverIo\Description\Api\Node\AbstractValueNode;
 use AppserverIo\Configuration\Interfaces\ValueInterface;
 
 /**