+++ src/AppserverIo/Appserver/Core/Api/Node/RewriteMapNode.php
@@ -20,6 +20,8 @@
 
 namespace AppserverIo\Appserver\Core\Api\Node;
 
+use AppserverIo\Description\Api\Node\AbstractNode;
+
 /**
  * Node class which represents the Access node of the configuration.
  *